Start Refactoring the Header Bar
