+++ frontend/src/common/redux/users-slice.js
@@ -1,5 +1,6 @@
 import { createSlice } from "@reduxjs/toolkit";
 
+// TODO Have something in here that handles login/logout? Set username was meant to do that though...
 export const usersSlice = createSlice({
   name: "users",
   initialState: {
